update delete funtion

+++ backend/controllers/stuff.js
@@ -1,4 +1,5 @@
 const Thing = require('../models/Thing');
+const fs = require('fs');
 
 exports.getOneThing = (req, res) => {
   Thing.findOne({ _id: req.params.id })
@@ -113,7 +114,20 @@ exports.modifyThing = (req, res) => {
 };
 
 exports.deleteThing = (req, res, next) => {
-  Thing.deleteOne({ _id: req.params.id })
-    .then(() => res.status(200).json({ message: 'Objet supprimé !'}))
-    .catch(error => res.status(400).json({ error }));
-}
+   Thing.findOne({ _id: req.params.id})
+       .then(thing => {
+           if (thing.userId != req.auth.userId) {
+               res.status(401).json({message: 'Not authorized'});
+           } else {
+               const filename = thing.imageUrl.split('/images/')[1];
+               fs.unlink(`images/${filename}`, () => {
+                   Thing.deleteOne({_id: req.params.id})
+                       .then(() => { res.status(200).json({message: 'Objet supprimé !'})})
+                       .catch(error => res.status(401).json({ error }));
+               });
+           }
+       })
+       .catch( error => {
+           res.status(500).json({ error });
+       });
+};
